69 convenience test
